+++ README.md
@@ -1,6 +1,6 @@
 # AI Snake Game
 
-This is a simple snake game written with Pygame. You can choose to play manually or let a basic AI control the snake. A main menu allows you to start the game, toggle AI, and adjust speed settings.
+This is a simple snake game written with Pygame. You can choose to play manually or let a basic AI control the snake. A main menu allows you to start the game, toggle AI, and adjust speed settings. The code is now split into a small `snake` package containing the game logic and menu helpers.
 
 ## Requirements
 - Python 3.12 or later
